--- packages/plugin-core/src/web/engine/INoteRenderer.ts
@@ -1,8 +1,0 @@
-import { RenderNoteOpts, RenderNoteResp } from "@dendronhq/common-all";
-
-/**
- * Extracted from DEngine
- */
-export interface INoteRenderer {
-  renderNote(opts: RenderNoteOpts): Promise<RenderNoteResp>;
-}
